--- a/rest/admin_api_auth_test.go
+++ b/rest/admin_api_auth_test.go
@@ -477,7 +477,6 @@ func TestAdminAuthWithX509(t *testing.T) {
 }
 
 func TestAdminAPIAuth(t *testing.T) {
-	base.LongRunningTest(t)
 
 	// Don't really care about the log level but this test hits the logging endpoint so this is used to reset the logging
 	base.SetUpTestLogging(t, base.LevelInfo, base.KeyNone)