--- a/rest/blip_api_replication_test.go
+++ b/rest/blip_api_replication_test.go
@@ -18,6 +18,8 @@ import (
 )
 
 func TestReplicationBroadcastTickerChange(t *testing.T) {
+	base.LongRunningTest(t)
+
 	if !base.TestUseXattrs() {
 		t.Skip("Skipping test that requires xattrs")
 	}
@@ -93,6 +95,8 @@ func TestReplicationBroadcastTickerChange(t *testing.T) {
 
 // TestBlipClientPushAndPullReplication sets up a bidi replication for a BlipTesterClient, writes documents on SG and the client and ensures they replicate.
 func TestBlipClientPushAndPullReplication(t *testing.T) {
+	base.LongRunningTest(t)
+
 	base.SetUpTestLogging(t, base.LevelTrace, base.KeyAll)
 
 	rtConfig := RestTesterConfig{