--- a/rest/bytes_read_public_api_test.go
+++ b/rest/bytes_read_public_api_test.go
@@ -24,6 +24,8 @@ import (
 )
 
 func TestBytesReadDocOperations(t *testing.T) {
+	base.LongRunningTest(t)
+
 	rt := NewRestTester(t, &RestTesterConfig{SyncFn: channels.DocChannelsSyncFunction})
 	defer rt.Close()