Refresh `LongRunningTest` list based on a new `gotestsum slowest` run for faster `-short` test suite runs.

## Methodology

```sh
$ gotestsum --jsonfile test-out.json -- -count=1 -p 1 ./...
$ gotestsum tool slowest --jsonfile test-out.json --threshold 250ms --skip-stmt='base.LongRunningTest(t)'
$ # and then some manual fixes and goimports
```

## Results
based on 250ms threshold
- All tests:
  - `DONE 4858 tests, 288 skipped in 448.183s`
- With `-short`:
  - `DONE 4026 tests, 532 skipped in 116.924s`